Move database.yml to template file
Move config/database.yml to config/database.yml.example as template
file.

Why?

Because config/database.yml is often modified on a per dev environment basis.

They need to be able to change this file, without having to untrack it
from git, and losing it when they git stash their untracked files.

Also:

I removed the "production" config and default "pool" option from the
template config/database.ymp.example.

Because if someone read that, they might think that this file will be
used in production, which it wont!